refactor(plugins): rename `test_filler` -> `filler`
