--- docs/library/pytest_plugins/test_filler.md
@@ -1,7 +1,0 @@
-# Test Filler Plugin
-
-::: pytest_plugins.test_filler
-
-::: pytest_plugins.test_filler.test_filler
-
-::: pytest_plugins.test_filler.pre_alloc
