+++ docs/navigation.md
@@ -44,4 +44,4 @@
     * [Ethereum Test VM Package](library/ethereum_test_vm.md)
     * [EVM Transition Tool Package](library/evm_transition_tool.md)
     * [Pytest Plugins](library/pytest_plugins/index.md)
-      * [Test Filler Plugin](library/pytest_plugins/test_filler.md)
+      * [Filler Plugin](library/pytest_plugins/filler.md)
