+++ pytest-consume.ini
@@ -6,4 +6,4 @@ addopts =
     -rxXs
     --tb short
     -p pytest_plugins.consume.consume
-    -p pytest_plugins.test_help.test_help
+    -p pytest_plugins.help.help
