+++ pytest-framework.ini
@@ -7,7 +7,6 @@ testpaths =
     src
 addopts = 
     -p pytester
-    --ignore=src/pytest_plugins/test_filler/test_filler.py
     --ignore=src/pytest_plugins/consume/direct/test_direct.py
     --ignore=src/pytest_plugins/consume/hive_simulators/engine/test_via_engine.py
     --ignore=src/pytest_plugins/consume/hive_simulators/rlp/test_via_rlp.py
