+++ pytest.ini
@@ -6,11 +6,11 @@ testpaths = tests/
 markers =
     slow
 addopts = 
-    -p pytest_plugins.test_filler.pre_alloc
-    -p pytest_plugins.test_filler.test_filler
+    -p pytest_plugins.filler.pre_alloc
+    -p pytest_plugins.filler.filler
     -p pytest_plugins.forks.forks
     -p pytest_plugins.spec_version_checker.spec_version_checker
-    -p pytest_plugins.test_help.test_help
+    -p pytest_plugins.help.help
     -m "not eip_version_check"
     --tb short
     --dist loadscope
